demo/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
websockets>=12
pydantic>=2
boto3>=1.34
python-dotenv>=1.0
httpx>=0.27
python-multipart>=0.0.9
